Hooked up some messages between gocInputPlayer and the new gocPhysicsSimple component.
Added a Font class to wrap font loading and text rendering.
Specialized the ToString in StringUtilities for floats, for nicer float rendering (mainly so the millisecond frametime counter wouldn't flicker about so much =/).
Renamed TestScreen to screenTest, and cleared a bunch of crap off it.
Rebuilt all the project files with premake.