Fixed a small bug in SFMLSound (wasn't new-ing an sf::SoundBuffer)
Moved RendererGL to a Renderers subdir.
Copied RendererGL to RendererGL3, started making it non-deprecated.
Moved SFMLGameWindow and the particle system to a GameWindows and ParticleSystem subdir, respectively.
Added a bunch of sound effects made with SFXR. Only one in use at the moment is bullet.wav (for testing).
Stripped more crap out of screenArenaShooter (formerly TestScreen).